# docs/swagger.yaml
definitions:
  dto.AdminInfoOutput:
    properties:
      avatar:
        type: string
      id:
        type: integer
      introduction:
        type: string
      login_time:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
  dto.AdminLoginInput:
    properties:
      password:
        example: "123456"
        type: string
      username:
        description: |-
          form 代表从请求中接收的字段
          todo validate换行的话会失效
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  dto.AdminLoginOutput:
    properties:
      token:
        example: token
        type: string
    type: object
  dto.ChangePwdInput:
    properties:
      password:
        example: "123456"
        type: string
    required:
    - password
    type: object
  dto.ServiceAddHTTPInput:
    properties:
      black_list:
        description: 黑名单ip
        type: string
      clientip_flow_limit:
        description: "\b客户端ip限流"
        type: integer
      header_transfor:
        description: header转换
        type: string
      ip_list:
        description: ip列表
        type: string
      need_https:
        description: 支持https
        type: integer
      need_strip_uri:
        description: 启用strip_uri
        type: integer
      need_websocket:
        description: 是否支持websocket
        type: integer
      open_auth:
        description: 关键词
        type: integer
      round_type:
        description: 轮询方式
        type: integer
      rule:
        description: 域名或者前缀
        type: string
      rule_type:
        description: 接入类型
        type: integer
      service_desc:
        description: 服务描述
        type: string
      service_flow_limit:
        description: 服务端限流
        type: integer
      service_name:
        description: 服务名
        type: string
      upstream_connect_timeout:
        description: 建立连接超时, 单位s
        type: integer
      upstream_header_timeout:
        description: 获取header超时, 单位s
        type: integer
      upstream_idle_timeout:
        description: 链接最大空闲时间, 单位s
        type: integer
      upstream_max_idle:
        description: 最大空闲链接数
        type: integer
      url_rewrite:
        description: url重写功能
        type: string
      weight_list:
        description: "\b权重列表"
        type: string
      white_list:
        description: 白名单ip
        type: string
    required:
    - ip_list
    - rule
    - service_desc
    - service_name
    - weight_list
    type: object
  dto.ServiceListItemOutput:
    properties:
      id:
        description: id
        type: integer
      load_type:
        description: 类型
        type: integer
      qpd:
        description: qpd
        type: integer
      qps:
        description: qps
        type: integer
      service_addr:
        description: 服务地址
        type: string
      service_desc:
        description: 服务描述
        type: string
      service_name:
        description: 服务名称
        type: string
      total_node:
        description: 节点数
        type: integer
    type: object
  dto.ServiceListOutput:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/dto.ServiceListItemOutput'
        type: array
      total:
        description: 总数
        type: integer
    type: object
  middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
info:
  contact: {}
paths:
  /admin/admin_info:
    get:
      consumes:
      - application/json
      description: 管理员信息
      operationId: /admin/admin_info
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminInfoOutput'
              type: object
      summary: 管理员信息
      tags:
      - 管理员接口
  /admin/change_pwd:
    post:
      consumes:
      - application/json
      description: 修改密码
      operationId: /admin/change_pwd
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePwdInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 修改密码
      tags:
      - 管理员接口
  /admin_login/login:
    post:
      consumes:
      - application/json
      description: 管理员登录
      operationId: /admin_login/login
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminLoginOutput'
              type: object
      summary: 管理员登录
      tags:
      - 管理员接口
  /admin_login/logout:
    get:
      consumes:
      - application/json
      description: 管理员退出
      operationId: /admin_login/logout
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 管理员退出
      tags:
      - 管理员接口
  /service/service_add_http:
    post:
      consumes:
      - application/json
      description: 添加HTTP服务
      operationId: /service/service_add_http
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddHTTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加HTTP服务
      tags:
      - 服务管理
  /service/service_delete:
    get:
      consumes:
      - application/json
      description: 服务删除
      operationId: /service/service_delete
      parameters:
      - description: 服务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 服务删除
      tags:
      - 服务管理
  /service/service_list:
    get:
      consumes:
      - application/json
      description: 服务列表
      operationId: /service/service_list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页个数
        in: query
        name: page_size
        required: true
        type: integer
      - description: 当前页数
        in: query
        name: page_no
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ServiceListOutput'
              type: object
      summary: 服务列表
      tags:
      - 服务管理
swagger: "2.0"